requirements.txt
aiosqlite
pyyaml
dataclasses-json
aiofiles
websockets
